requirements.txt
alembic==1.4.2
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cymem==2.0.3
dnspython==2.0.0
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
Mako==1.1.3
MarkupSafe==1.1.1
murmurhash==1.0.2
numpy==1.19.1
pandas==1.1.1
plac==1.1.3
preshed==3.0.2
psycopg2-binary==2.8.5
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
spacy==2.3.2
SQLAlchemy==1.3.19
srsly==1.0.2
thinc==7.4.1
threadpoolctl==2.1.0
tqdm==4.48.2
urllib3==1.25.10
wasabi==0.7.1
Werkzeug==1.0.1
zipp==3.1.0
